+++ pywal/backends/colorz.py
@@ -2,9 +2,15 @@
 Generate a colorscheme using Colorz.
 """
 import logging
-import shutil
-import subprocess
 import sys
+
+try:
+    import colorz
+
+except ImportError:
+    logging.error("colorz wasn't found on your system.")
+    logging.error("Try another backend. (wal --backend)")
+    sys.exit(1)
 
 from .. import colors
 from .. import util
@@ -12,26 +18,26 @@ from .. import util
 
 def gen_colors(img):
     """Generate a colorscheme using Colorz."""
-    cmd = ["colorz", "-n", "6", "--bold", "0", "--no-preview", "--no-bg-img"]
-    return subprocess.check_output([*cmd, img]).splitlines()
+    # pylint: disable=not-callable
+    raw_colors = colorz.colorz(img, n=6, bold_add=0)
+    return [util.rgb_to_hex([*color[0]]) for color in raw_colors]
 
 
 def adjust(cols, light):
     """Create palette."""
-    bg = util.blend_color("#555555", cols[1])
-
-    raw_colors = [bg, *cols, "#FFFFFF",
-                  "#333333", *cols, "#FFFFFF"]
+    raw_colors = [cols[0], *cols, "#FFFFFF",
+                  "#000000", *cols, "#FFFFFF"]
 
     return colors.generic_adjust(raw_colors, light)
 
 
 def get(img, light=False):
     """Get colorscheme."""
-    if not shutil.which("colorz"):
-        logging.error("Colorz wasn't found on your system.")
-        logging.error("Try another backend. (wal --backend)")
+    cols = gen_colors(img)
+
+    if len(cols) < 6:
+        logging.error("colorz failed to generate enough colors.")
+        logging.error("Try another backend or another image. (wal --backend)")
         sys.exit(1)
 
-    cols = [col.decode('UTF-8').split()[0] for col in gen_colors(img)]
     return adjust(cols, light)